refactor: cmp
