+++ frontend/src/lib/components/accounts/IcrcAccountsPage.svelte
